server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
youtube-transcript-api==1.2.4
mecab-python3==1.0.9
mecab-ko-dic==1.0.0
pydantic==2.10.4
